requirements.txt
discord.py==1.6.0
python-dotenv==0.15.0
jishaku==1.20.0.220
requests==2.23.0
firebase-admin==4.5.1